% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_rejection_frequency}
\alias{get_rejection_frequency}
\title{compute rejection frequencies}
\usage{
get_rejection_frequency(n_sims, G, k, rho, gamma, workers)
}
\description{
compute rejection frequencies
}
\examples{
get_rejection_frequency(n_sims = 10, G = 10, k = 9, rho = 0.1, gamma = 2, workers = 1)
}
